backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
tensorflow>=2.20.0
pillow>=10.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
python-multipart>=0.0.6
pydantic>=2.5.0
numpy>=1.26.0
psutil>=5.9.0
pandas>=2.0.0
scikit-learn>=1.3.0